Developer friendly solution with NodeJs and Web3
This is an example of inserting a string and the smart contract take
the string, sha256 it, mapping it to an integer and store in the
blockchain. And then return to the console what the sha256 is.

It is simple, but, these is no remix need for getting Abi or Byte code,
the code will generate it automatically. Also, template provide ability
to select accounts in testrpc.